Updated on 2025-11-26

+++ app/src/main/java/com/tangem/data/network/model/BinanceFee.kt
@@ -12,5 +12,5 @@ data class BinanceFixedFee(
         var msg_type: String? = null,
 
         @SerializedName("fee")
-        var fee: Long? = null
+        var fee: Int? = null
 )
